third_party/WebKit/LayoutTests/fast/dynamic/first-child-descendant.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<style>
#b { color: green; }
#a:first-child #b { color: red }
</style>
<div id="t">
    <div id="a">
        <div id="b">This text should be green</div>
    </div>
</div>
<script>
description("Evaluation of :first-child affecting descendant.");

t.offsetTop;
a.style.backgroundColor = "white";
t.offsetTop;
t.insertBefore(document.createElement("div"), a);

var green = "rgb(0, 128, 0)";

shouldBe("getComputedStyle(b).color", "green");
</script>
